Shutup Clang about reserved macros. Also fixing issue with Blip_Buffer using floor instead of floorf.

+++ mednafen/sound/Blip_Buffer.c
@@ -105,7 +105,7 @@ blargg_err_t Blip_Buffer_set_sample_rate(Blip_Buffer* bbuf, long new_rate, int m
 blip_resampled_time_t Blip_Buffer_clock_rate_factor(Blip_Buffer* bbuf, long rate)
 {
 	float ratio = (float) bbuf->sample_rate / rate;
-	blip_s64 factor = (blip_s64) floor(ratio * (1LL << BLIP_BUFFER_ACCURACY) + 0.5f);
+	blip_s64 factor = (blip_s64) floorf(ratio * (1LL << BLIP_BUFFER_ACCURACY) + 0.5f);
 	assert(factor > 0 || !bbuf->sample_rate);   // fails if clock/output ratio is too large
 	return (blip_resampled_time_t) factor;
 }
